+++ Modules/AntiAI/AntiAI.psm1
@@ -11,7 +11,7 @@
     
 .NOTES
     Module: AntiAI
-    Version: 2.2.0
+    Version: 2.2.1
     Author: NoID Privacy
 #>
 
@@ -29,7 +29,7 @@ $privateFunctions = @(
     'Disable-Recall'
     'Set-RecallProtection'
     'Disable-Copilot'
-    'Disable-CopilotAdvanced'  # NEW v2.2.0: URI handlers, Edge sidebar, Recall export
+    'Disable-CopilotAdvanced'  # NEW v2.2.1: URI handlers, Edge sidebar, Recall export
     'Disable-ClickToDo'
     'Disable-SettingsAgent'
     'Disable-ExplorerAI'       # NEW: File Explorer AI Actions menu