+++ Modules/DNS/DNS.psm1
@@ -12,7 +12,7 @@
     
 .NOTES
     Author: NoID Privacy
-    Version: 2.2.0
+    Version: 2.2.1
     Requires: PowerShell 5.1+, Administrator privileges
 #>